+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.10 2024-11-??
-* Add SI1151 sunlight sensor
+
+* Add SI1145 sunlight sensor
 * Rework README
 
 ## 0.9.9 2024-11-17